# Conflicts:
#	.idea/dataSources.xml
#	src/main/java/eus/ehu/ridesfx/businessLogic/BlFacade.java
#	src/main/java/eus/ehu/ridesfx/businessLogic/BlFacadeImplementation.java
#	src/main/java/eus/ehu/ridesfx/dataAccess/DataAccess.java
#	src/main/java/eus/ehu/ridesfx/domain/Traveler.java
#	src/main/java/eus/ehu/ridesfx/uicontrollers/QueryRidesController.java
#	src/main/resources/eus/ehu/ridesfx/ui/QueryRides.fxml
#	src/main/resources/hibernate.cfg.xml

+++ src/main/java/eus/ehu/ridesfx/domain/Ride.java
@@ -165,8 +165,12 @@ public class Ride implements Serializable {
 		this.date = date;
 	}
 
-
-	public float getNumPlaces() {
+	/**
+	 * Get the free places of the ride
+	 *
+	 * @return the available places
+	 */
+	public int getNumPlaces() {
 		return numPlaces;
 	}
